APOLLO-21277: refactored fields component
