+++ java-sdk/connectors/security-filtering-connector/src/main/java/com/lucidworks/connectors/plugins/security/component/ExampleSecurityFilterFieldsComponent.java
@@ -1,8 +1,6 @@
-package com.lucidworks.fusion.connector.plugin.validation;
+package com.lucidworks.connectors.plugins.security.component;
 
-import com.lucidworks.fusion.connector.plugin.api.security.SecurityFilter;
 import com.lucidworks.fusion.connector.plugin.api.security.SecurityFilterQueryFieldsComponent;
-import com.lucidworks.fusion.connector.plugin.api.security.Subject;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -17,8 +15,4 @@ public class ExampleSecurityFilterFieldsComponent implements SecurityFilterQuery
     return securityFilterFields;
   }
 
-  @Override
-  public SecurityFilter buildSecurityFilter(Subject subject) {
-    throw new UnsupportedOperationException("buildSecurityFilter is unsupported");
-  }
 }
